% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{aligned_pmax}
\alias{aligned_pmax}
\title{Version of pmax which doesn't recycle, but instead extends each vector}
\usage{
aligned_pmax(...)
}
\arguments{
\item{...}{multiple numeric vector arguments}
}
\value{
pmax across all vector without recycling
}
\description{
Version of pmax which doesn't recycle, but instead extends each vector
}
