% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\name{ansi}
\alias{ansi}
\title{Colour a string with ANSI}
\usage{
ansi(x, fg = NULL, bg = NULL, sep = "", use_ansi)
}
\arguments{
\item{x}{character string}

\item{fg, bg}{foreground and background colours. NULL for nothing}

\item{sep}{seperator for collapseing strings.}

\item{use_ansi}{default: TRUE}
}
\description{
Colour a string with ANSI
}
