% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terse.R
\name{build_names}
\alias{build_names}
\title{Build character strings to represent the names}
\usage{
build_names(x, width = 10, config = list())
}
\arguments{
\item{x}{an object with names. Usually list or data.frame}

\item{width}{how wide to render the name itself?}

\item{config}{named list of user configuration options}
}
\description{
Build character strings to represent the names
}
