% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-frame.R
\name{calc_max_element_widths}
\alias{calc_max_element_widths}
\title{Calculate the maximum element width for each row of a data.frame}
\usage{
calc_max_element_widths(df)
}
\arguments{
\item{df}{data.frame}
}
\description{
Calculate the maximum element width for each row of a data.frame
}
