% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\name{cansi}
\alias{cansi}
\title{Colour and output a string with ANSI}
\usage{
cansi(x, fg = NULL, bg = NULL, sep = "", use_ansi = TRUE)
}
\arguments{
\item{x}{character string}

\item{fg}{foreground and background colours. NULL for nothing}

\item{bg}{foreground and background colours. NULL for nothing}

\item{sep}{seperator for collapseing strings.}

\item{use_ansi}{default: TRUE}
}
\description{
Colour and output a string with ANSI
}
