% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\name{colour_vec}
\alias{colour_vec}
\title{Colour a vector using a highlight for every nth element}
\usage{
colour_vec(vec, n, colour_all, colour_nth, use_ansi)
}
\arguments{
\item{vec}{atomic which can be coerced to character}

\item{n}{the interval to highlight. Set to 0 to disable highlighting}

\item{colour_all}{ansi colour of all elements}

\item{colour_nth}{ansi colour applied every nth element}

\item{use_ansi}{default: TRUE}
}
\description{
Colour a vector using a highlight for every nth element
}
