% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terse.R
\name{ellipsis}
\alias{ellipsis}
\title{If string exceeds with, truncate with "..." - ANSI aware.}
\usage{
ellipsis(text, width = 0)
}
\arguments{
\item{text}{character string}

\item{width}{default: 0 (auto). Set to negative for no truncation.}
}
\description{
If string exceeds with, truncate with "..." - ANSI aware.
}
