% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{extend_to_length}
\alias{extend_to_length}
\title{Extend a vector to the given length by appending the given value}
\usage{
extend_to_length(vec, len, value = -Inf)
}
\arguments{
\item{vec}{numeric vector}

\item{len}{desired length}

\item{value}{value to append to shorter vectors}
}
\value{
vector of the requested length
}
\description{
Extend a vector to the given length by appending the given value
}
