% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-frame.R
\name{get_element_widths_vec}
\alias{get_element_widths_vec}
\title{Calculate the width of the character representation of each element in a vector}
\usage{
get_element_widths_vec(vec)
}
\arguments{
\item{vec}{presumed atomic vector which is a column in a data.frame}
}
\description{
If the given vec is not actually atomic, just pretend it has a width of 0.
}
\details{
Treat NAs as 2 characters wide
}
