% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_deparse}
\alias{long_deparse}
\title{Deparse an object to a single string}
\usage{
long_deparse(x)
}
\arguments{
\item{x}{object}
}
\value{
single character string
}
\description{
base::deparse has a maximum width of 500 characters before breaking into
multiple strings. This wrapper around deparse just concatenates these
multiple strings
}
