% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{max_length}
\alias{max_length}
\title{Determine maximum lengths of any object anywhere in list or data.frame}
\usage{
max_length(x)
}
\arguments{
\item{x}{object, usually a list or data.frame}
}
\value{
maximum element length
}
\description{
Determine maximum lengths of any object anywhere in list or data.frame
}
