% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{max_name_length}
\alias{max_name_length}
\title{Determine maximum name lengths at all depths}
\usage{
max_name_length(x)
}
\arguments{
\item{x}{object, usually a list or data.frame}
}
\value{
vector of the maximum name length at each depth
}
\description{
Determine maximum name lengths at all depths
}
