% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\name{nchar_ansi}
\alias{nchar_ansi}
\title{Remove ansi codes from a string and calculate nchar()}
\usage{
nchar_ansi(x)
}
\arguments{
\item{x}{character string}
}
\description{
Remove ansi codes from a string and calculate nchar()
}
