% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_to_width}
\alias{pad_to_width}
\title{Pad strings to a given width}
\usage{
pad_to_width(x, len, align = c("l", "r"))
}
\arguments{
\item{x}{character strings}

\item{len}{length}

\item{align}{align text to the left or right? default: left}
}
\description{
Pad strings to a given width
}
