% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{padding}
\alias{padding}
\title{Generate a character vector of padding strings}
\usage{
padding(lens)
}
\arguments{
\item{lens}{lengths}
}
\value{
vector of character strings containing blanks of the given lengths
}
\description{
Generate a character vector of padding strings
}
