% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\docType{data}
\name{reset}
\alias{reset}
\title{Reset back to terminal defaults}
\format{
An object of class \code{character} of length 1.
}
\usage{
reset
}
\description{
Reset back to terminal defaults
}
\keyword{datasets}
