% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terse.R
\name{terse}
\alias{terse}
\alias{terse.default}
\alias{terse.data.frame}
\alias{terse.list}
\alias{terse.defaultlist}
\alias{terse.lm}
\title{Output compact representation of R objects}
\usage{
terse(
  x,
  prefix = "",
  width = 0,
  name_widths = NULL,
  element_widths = NULL,
  max_vec_len = NULL,
  config = list()
)

\method{terse}{default}(
  x,
  prefix = NULL,
  width = 0,
  name_widths = NULL,
  element_widths = NULL,
  max_vec_len = 100,
  config = list()
)

\method{terse}{data.frame}(
  x,
  prefix = NULL,
  width = 0,
  name_widths = NULL,
  element_widths = NULL,
  max_vec_len = NULL,
  config = list(),
  ...
)

\method{terse}{list}(
  x,
  prefix = NULL,
  width = 0,
  name_widths = NULL,
  element_widths = NULL,
  max_vec_len = NULL,
  config = list(),
  ...
)

\method{terse}{defaultlist}(
  x,
  prefix = NULL,
  width = 0,
  name_widths = NULL,
  element_widths = NULL,
  max_vec_len = NULL,
  config = list(),
  ...
)

\method{terse}{lm}(
  x,
  prefix = NULL,
  width = 0,
  name_widths = NULL,
  element_widths = NULL,
  max_vec_len = NULL,
  config = list(),
  ...
)
}
\arguments{
\item{x}{R object}

\item{prefix}{prefix}

\item{width}{target width. 0 = auto. -1 = no limit}

\item{name_widths}{vector of name widths}

\item{element_widths}{widths to print each individual element in a vector.
default: NULL (natural width), single value to apply to all elements,
or a numeric vector the same length as x}

\item{max_vec_len}{maximum vector length anywhere in original object}

\item{config}{named list of user configuration options
\itemize{
\item{\code{ansi       } - Use ANSI to colour output. default: TRUE}
\item{\code{soft       } - non-highlight colour. default: grey40}
\item{\code{gsep       } - separator for vector output. default: ','}
\item{\code{colour_nth } - colour for every nth vector element. default: blue4}
\item{\code{colour_all } - colour for all other vector elements. default: grey40}
\item{\code{colour_type} - colour for the type/length meta information. default: seagreen}
\item{\code{nth        } - interval for colouring vector elements. default: 0 (off)}
}}

\item{...}{other arguments}
}
\description{
Output compact representation of R objects
}
