% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terse.R
\name{terse_core}
\alias{terse_core}
\title{Output a single line representation of an object}
\usage{
terse_core(
  single_line,
  vec_type,
  vec_len,
  prefix = NULL,
  width = 0,
  name_widths = NULL,
  max_vec_len = 99,
  config = list()
)
}
\arguments{
\item{single_line}{Representation of R object in a single line}

\item{vec_type}{single character to represent class of object}

\item{vec_len}{length of this object}

\item{prefix}{prefix}

\item{width}{target width. 0 = auto. -1 = no limit}

\item{name_widths}{vector of name widths}

\item{max_vec_len}{maximum length of all objects to be be presented.}

\item{config}{user config options in a named list}
}
\description{
Output a single line representation of an object
}
