% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Basic version of modifyList stolen from ggplot2}
\usage{
update_list(old, new)
}
\arguments{
\item{old, new}{named lists}
}
\description{
Basic version of modifyList stolen from ggplot2
}
